+++ WorldCities/WorldCities/src/app/app.component.ts
@@ -8,8 +8,10 @@ import { Component, OnInit } from '@angular/core';
   styleUrl: './app.component.scss'
 })
 export class AppComponent implements OnInit {
-  ngOnInit(): void {
-      throw new Error('Method not implemented.');
-  }
   title = 'WorldCities';
+
+  ngOnInit(): void {
+ 
+  }
+
 }