+++ WorldCities/WorldCities/src/app/nav-menu/nav-menu.component.ts
@@ -1,9 +1,12 @@
-import { Component } from '@angular/core';
+import { Component, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 
 @Component({
   selector: 'app-nav-menu',
   templateUrl: './nav-menu.component.html',
-  styleUrl: './nav-menu.component.scss'
+  styleUrl: './nav-menu.component.scss',
+  schemas: [
+    CUSTOM_ELEMENTS_SCHEMA
+  ]
 })
 export class NavMenuComponent {